feat(mps-sync-plugin-lib): refactoring move classes to new packages
