+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/mps/factories/SolutionProducer.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.neu
+package org.modelix.mps.sync.mps.factories
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.util.io.systemIndependentPath
@@ -31,6 +31,7 @@ import jetbrains.mps.smodel.GeneralModuleFactory
 import jetbrains.mps.vfs.IFile
 import jetbrains.mps.vfs.VFSManager
 import org.modelix.kotlin.utils.UnstableModelixFeature
+import org.modelix.mps.sync.mps.util.runWriteInEDTBlocking
 import java.io.File
 
 // TODO hacky solution. A nicer one could be: https://github.com/JetBrains/MPS/blob/master/workbench/mps-platform/jetbrains.mps.ide.platform/source_gen/jetbrains/mps/project/modules/SolutionProducer.java
