+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/mps/listeners/ModelChangeListener.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.neu.listeners
+package org.modelix.mps.sync.mps.listeners
 
 import jetbrains.mps.smodel.SModelInternal
 import jetbrains.mps.smodel.event.SModelChildEvent
@@ -39,7 +39,7 @@ import org.modelix.model.api.getNode
 import org.modelix.model.mpsadapters.MPSChildLink
 import org.modelix.model.mpsadapters.MPSConcept
 import org.modelix.model.mpsadapters.MPSReferenceLink
-import org.modelix.mps.sync.neu.MpsToModelixMap
+import org.modelix.mps.sync.transformation.MpsToModelixMap
 import org.modelix.mps.sync.util.nodeIdAsLong
 import org.modelix.mps.sync.util.runIfAlone
 import java.util.concurrent.atomic.AtomicReference
