+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/transformation/MpsToModelixMap.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.neu
+package org.modelix.mps.sync.transformation
 
 import org.jetbrains.mps.openapi.model.SModel
 import org.jetbrains.mps.openapi.model.SModelId
